server/pyproject.toml
[project]
name = "trailine-server"
version = "0.0.1"
description = "Hiking Guilde Service Server"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django-environ>=0.12.0",
    "django-filter>=25.1",
    "django-redis>=6.0.0",
    "djangorestframework>=3.16.0",
    "djangorestframework-camel-case>=1.4.2",
    "djangorestframework-simplejwt>=5.5.0",
    "drf-yasg>=1.21.10",
    "markdown>=3.8",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "factory-boy>=3.3.3",
]
